refactor(ai): extract OpenAI chat completion base provider
